Align config struct

+++ main/config.h
@@ -18,7 +18,7 @@ struct map_conf {
 
 struct config {
     uint32_t magic;
-    uint8_t multitap_conf;
+    uint32_t multitap_conf;
     uint8_t dev_mode[WIRED_MAX_DEV];
     uint8_t map_size[WIRED_MAX_DEV];
     struct map_conf map_conf[WIRED_MAX_DEV][ADAPTER_MAPPING_MAX];
